Implemented frame-independent entity animation

+++ src/timers/fps-display-timer.cpp
@@ -4,7 +4,7 @@
 
 
 void FPSDisplayTimer::calculateFPS() {
-    mAverageFPS = mAccumulatedFrames / (static_cast<int>(getTicks()) / 1000.0f);
+    mAverageFPS = mAccumulatedFrames / (getTicks() / 1000.0f);
     if (mAccumulatedFrames >= std::numeric_limits<unsigned long long int>::max()) mAccumulatedFrames = 0;
     ++mAccumulatedFrames;
 }
